+++ src/pages/PostsManagerPage.tsx
@@ -1,13 +1,14 @@
 import { useEffect, useState } from "react"
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
 import { useLocation, useNavigate } from "react-router-dom"
-import { Button } from "../shared/button/Button"
+import { Button } from "../shared/ui/button/Button"
+import { Card, CardContent, CardHeader, CardTitle } from "../shared/ui/card/Card"
 import {
   // Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
+  // Card,
+  // CardContent,
+  // CardHeader,
+  // CardTitle,
   Dialog,
   DialogContent,
   DialogHeader,
